Use ubuntu-20.04 instead of latest in github actions
ubuntu-latest is currently default to ubuntu-18.04
which does not contain Java 11 but Java 8 only.

Because of this, it is needed to use the setup-java
to download a JDK 11.

This can be avoided by forcing use of ubuntu-20.04.

Once ubuntu-latest points to 20.04, switch can be done
back to ubuntu-latest.

See https://github.com/actions/virtual-environments/issues/1816