Remove native image and rely on the JVM at runtime
The complexity of setting a build pipeline for GraalVM native image
and having access to both macOs/linux for both x64 and ARM architecture
is too much of an overhead for a small opensource project.

This might be reconsidered in the future when availability on those
OS/arch will be more widely spread or that GraalVM will offer
cross compilation support.

See https://github.com/oracle/graal/issues/407