Revert "Add support for building with java early access versions"
This reverts commit 543111d29abca4961c96346c38882e154e82eec2.
Reason is that the JVM toolchain is specified in the gradle build.

Hence, changing the java version in the pipeline only affects
the gradle wrapper, but not the compilation nor the tests.

Also, this creates non determinism about the upload test results task
that is part of the matrix run.

Hence, disabling for now.