build: Introduce maven pom
It turns out that integrating symly into
native package managers is a difficult exercise
due to the complexity of the Gradle packaging.

This is an attempt to make a working maven build
prototype. If proven successful, this might
replace the existing gradle build. If not,
it might coexist in order to enable the
build in debian or fedora package managers.